splunk-mcp-server2/python/pyproject.toml
[project]
name = "splunk-mcp-server"
version = "0.1.0"
description = "MCP server for retrieving data from Splunk"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.3.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
]

[project.scripts]
splunk-mcp-server = "server:main"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["server", "helpers", "splunk_client"]